(룸메이트 게시판) 페이징 카테고리별 분류

+++ room/src/main/java/com/room/controller/MateBoardController.java
@@ -1,7 +1,5 @@
 package com.room.controller;
 
-import java.io.File;
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
@@ -11,18 +9,13 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.View;
-import org.springframework.web.servlet.view.RedirectView;
 
-import com.room.dto.FBoard;
 import com.room.dto.MateBoard;
 import com.room.service.MateBoardService;
-import com.room.service.MateBoardServiceImpl;
+import com.room.ui.JinPager;
 
 @Controller
 @RequestMapping(path= {"/mate-board"})
@@ -32,11 +25,23 @@ public class MateBoardController {
 	@Qualifier("mateBoardService")
 	private MateBoardService mateBoardService;
 
-	@GetMapping(path = { "/","/list" })
-	public String list(Model model) {
+	@GetMapping(path = { "/list" })
+	public String list(@RequestParam(defaultValue = "1")int pageNo,
+			Model model) {
 		
-		List<MateBoard> mateBoardList = mateBoardService.findAll();
+		int pageSize = 3;
+		int pagerSize = 3;
+		int count = 0;
+		
+		List<MateBoard> mateBoardList = mateBoardService.findByPage(pageNo, pageSize);
+		count = mateBoardService.findBoardCount("roommate");
+		
+		JinPager pager = new JinPager(count, pageNo, pageSize, pagerSize, "list");
+		
 		model.addAttribute("mateBoardList", mateBoardList);
+		model.addAttribute("pager", pager);
+		model.addAttribute("pageNo", pageNo);
+		
 		
 		return "mate-board/list";  // --> /WEB-INF/views/ + board/list + .jsp
 	}
@@ -53,32 +58,36 @@ public class MateBoardController {
 		
 		mateBoardService.writeBoard(board);			
 		
-		 return "redirect:/mate-board/list";
+		 return "redirect:list";
 	}
 	
 	@GetMapping(path = {"/detail"})
 	public String detail(@RequestParam(name="boardNo" , defaultValue = "-1")int boardNo,
+						@RequestParam(name="pageNo", defaultValue = "-1")int pageNo,
 						Model model) {
-		if(boardNo == -1) {
+		if(boardNo == -1 || pageNo == -1) {
 			return "redirect:list";
 		}
 		
 		MateBoard board = mateBoardService.findByBoardNo(boardNo);
 		if(board == null) { // 해당 번호의 게시글이 없는 경우
-			System.out.println(boardNo);
 			return "redirect:list";
 		}
 		
-		model.addAttribute("board",board);
+		model.addAttribute("board", board);
+		model.addAttribute("pageNo", pageNo);
 		
 		return "mate-board/detail";
 	}
 	
 	@GetMapping(path= {"/delete"})
 	public String delete(
-			@RequestParam(name = "boardNo", defaultValue = "-1")int boardNo) {
-		if (boardNo > 0) {
+			@RequestParam(name = "boardNo", defaultValue = "-1")int boardNo,
+			@RequestParam(defaultValue = "-1")int pageNo) {
+		
+		if (boardNo > 0 && pageNo > 0) {
 			mateBoardService.delete(boardNo);
+			return "redirect:list?pageNo=" + pageNo;
 		}
 		return "redirect:list";
 	}
@@ -86,9 +95,10 @@ public class MateBoardController {
 	@GetMapping(path = {"/edit"})
 	public String showEditForm(
 			@RequestParam(name="boardNo", defaultValue = "-1")int boardNo,
-			Model model){
+			@RequestParam(defaultValue = "-1")int pageNo,
+			Model model) {
 			
-			if(boardNo < 1) {
+			if(boardNo < 1 && pageNo < 1) {
 				return "redirect:list";
 			}
 			
@@ -98,6 +108,7 @@ public class MateBoardController {
 			}
 			
 			model.addAttribute("board",board);
+			model.addAttribute("pageNo", pageNo);
 			
 			return "mate-board/edit";
 				
@@ -105,11 +116,16 @@ public class MateBoardController {
 	
 	
 	@PostMapping(path = {"/edit"})
-	public String edit(MateBoard board) {
+	public String edit(MateBoard board,
+						@RequestParam(defaultValue = "-1")int pageNo) {
+		
+		if (pageNo < 1) {
+			return "redirect:list";
+		}
 		
 		mateBoardService.update(board);
 		
-		return String.format("redirect:detail?boardNo=%d", board.getBoardNo());
+		return String.format("redirect:detail?boardNo=%d", board.getBoardNo(), pageNo);
 		
 	}
